Add type definitions
This adds basic type definitions for the exported `quantile` function
as well as the `options` object it takes as a parameter. This will
make it wasier for this package to be used in typescript projects.

The `Matrix` type is not included in the type definitions, as there
appear to be no type definitions for it and it is part of a separate
package.